move settings and structs to files
